apps/jobsrapper/pyproject.toml
[project]
name = "job-hunter-sentinel"
version = "0.1.0"
description = "AI-powered job scraping and recommendation system for O-1 visa applicants"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Resume Matcher Team" }
]

dependencies = [
    "python-jobspy>=1.1.77",
    "google-generativeai>=0.3.0",
    "pandas>=2.0.0",
    "resend>=0.7.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py310', 'py311']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
